requirements.txt
langchain
langchain-openai
openai-whisper
sounddevice
pyttsx3
